备份

+++ bsp/tracker.h
@@ -7,7 +7,7 @@
 
 #include <stdbool.h>
 #include "main.h"
-int get_tracker_num(void);
+float get_tracker_num(void);
 extern bool tracker_on;
 void tracker_set(bool if_on);
 #endif //HOLTEK_TRACKER_H
